move server to server_helpers_test.go
